requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
PyJWT==1.7.1
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
